Main modular structure

+++ src/server.c
@@ -13,53 +13,38 @@
 
 static int public_fifo_fd = -1;
 
-int main(){
-    /*Testing values*/
-    int nsecs = 5;
-    char name[20] = "/tmp/a_fifo";
-    unsigned buf_size = 10;
-
-    struct timespec remaining_time;
-
-    if (setup_timer(nsecs) == -1) {
-        exit(EXIT_FAILURE);
-    }
-
+void make_fifo(char* name){
     if(mkfifo(name, 0666) != 0){
         if(errno != EEXIST){
             fprintf(stderr, "Could not creat the fifo\n");
             exit(EXIT_FAILURE);
         }
     }
+}
 
-    pthread_t id;
-    pthread_attr_t tatrr;
-    if (pthread_attr_init(&tatrr) != 0) {
+int initialize_pthread(pthread_attr_t* tatrr){
+    if (pthread_attr_init(tatrr) != 0) {
         perror("Could not initialize pthread attribute");
         return -1;
     }
-    if (pthread_attr_setdetachstate(&tatrr, PTHREAD_CREATE_DETACHED) != 0) {
+    if (pthread_attr_setdetachstate(tatrr, PTHREAD_CREATE_DETACHED) != 0) {
         perror("Could not set pthread attribute");
-        pthread_attr_destroy(&tatrr);
+        pthread_attr_destroy(tatrr);
         return -1;
     }
+    return 0;
+}
 
-    if(init_producer_consumer(buf_size) != 0){
-        fprintf(stderr, "Could not init producer consumer\n");
-        exit(EXIT_FAILURE);
-    }
-
-    public_fifo_fd = open(name, O_RDONLY);
-    if(public_fifo_fd == -1){
-        fprintf(stderr, "Could not open fifo\n");
-        exit(EXIT_FAILURE);
-    } 
-
-    if (pthread_create(&id, &tatrr, consumer, NULL) != 0) {
+void creat_consumer( pthread_t* id, pthread_attr_t* tatrr){
+    if (pthread_create(id, tatrr, consumer, NULL) != 0) {
         perror("Could not create thread");
         exit(EXIT_FAILURE);
     }
-    
+
+}
+
+void listener( pthread_t* id, pthread_attr_t* tatrr){
+    struct timespec remaining_time;
 
     while(true){
         if (get_timer_remaining_time(&remaining_time) == -1) {
@@ -77,19 +62,61 @@ int main(){
         if(message != NULL){
             log_operation(RECVD, message->rid, message->tskload, message->tskres); //maybe this is not the the answer but that will do for test
             push_pending_request(message);
-            if (pthread_create(&id, &tatrr, producer, NULL) != 0) {
+            if (pthread_create(id, tatrr, producer, NULL) != 0) {
                 perror("Could not create producer thread");
             }
         }
     }
+}
 
-    destroy_timer();
+void open_fifo(char* name){
+    public_fifo_fd = open(name, O_RDONLY);
+    if(public_fifo_fd == -1){
+        fprintf(stderr, "Could not open fifo\n");
+        exit(EXIT_FAILURE);
+    } 
+}
 
+void close_fifo(){
     if(close(public_fifo_fd) != 0){
         perror("Could not close the fifo\n");
     }
+}
+
+int main(){
+    /*Testing values*/
+    int nsecs = 5;
+    char name[20] = "/tmp/a_fifo";
+    unsigned buf_size = 10;
+
+    if (setup_timer(nsecs) == -1) {
+        exit(EXIT_FAILURE);
+    }
+
+    make_fifo(name);
+
+    pthread_t id;
+    pthread_attr_t tatrr;
+
+    initialize_pthread(&tatrr);
+
+    if(init_producer_consumer(buf_size) != 0){
+        fprintf(stderr, "Could not init producer consumer\n");
+        exit(EXIT_FAILURE);
+    }
+
+    open_fifo(name);
+
+    creat_consumer(&id, &tatrr);
+
+    listener(&id, &tatrr);
+
+    destroy_timer();
+
+    close_fifo();
 
     atexit(destroy_producer_consumer);
+
     pthread_exit(NULL);
 
     /*MessageQueue* queue = init_message_queue();
